rag_service-open/rag_service/requirements.txt
llama-index
llama-index-llms-openrouter
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-embeddings-huggingface
llama-index-readers-file
chromadb
GitPython
python-dotenv
